Better anti-flapping filtering
We are queuing tasks for events within a given time window. The process
is delayed to the end of the window to avoid to delete/recreate DNS
entries when a container is restarted. If events are comming during our
process, we are queuing a new process window (before we had a race
condition that leads to lost events).